RichardBurton.Codec (richard_burton v0.1.0)

Utilities for serialization and deserialization

Functions

flatten(map)

Flattens a nested map, post-processing the resulting {key, value} pairs with the given post function. Lists are not traversed. Keys on nested maps are regenerated prepending the key of their parent map.

Returns the flattened map with string keys.

## Examples

  iex> RichardBurton.Codec.flatten(%{})
  %{}

  iex> RichardBurton.Codec.flatten(%{a: 1, b: 2, c: 3})
  %{"a" => 1, "b" => 2, "c" => 3}

  iex> RichardBurton.Codec.flatten(%{a: [1, 2, %{a: 1}]})
  %{"a" => [1, 2, %{a: 1}]}

  iex> RichardBurton.Codec.flatten(%{a: 1, b: %{a: 1, b: 2, c: 3}})
  %{"a" => 1, "b_a" => 1, "b_b" => 2, "b_c" => 3}

  iex> RichardBurton.Codec.flatten(%{a: 1, b: [2, 3, %{a: 1, b: 2}], c: %{a: 1, b: 2, c: 3}})
  %{ "a" => 1, "b" => [2, 3, %{a: 1, b: 2}], "c_a" => 1, "c_b"=> 2, "c_c"=> 3}

nest(map)

Automatically nests a flat map. Keys are grouped by their Longest Common Prefix (LCP) to define nesting points. LCP is calculated considering atomic words separated by underscore:

`aBC` and `aDE do not have a common prefix
`a_BC` and `a_DE` have "a" as common prefix
`a_b_BC` and `a_b_DE` have "a_b" as common prefix

Returns the nested map with string keys.

## Examples

  iex> RichardBurton.Codec.nest(%{})
  %{}

  iex> RichardBurton.Codec.nest(%{a: 1, b: 2, c: 3})
  %{"a" => 1, "b" => 2, "c" => 3}

  iex> RichardBurton.Codec.nest(%{a: 1, b_a: 1, b_b: 2, b_c: 3})
  %{"a" => 1, "b" => %{"a" => 1, "b" => 2, "c" => 3}}

  iex> RichardBurton.Codec.nest(%{a: 1, b_a: 1, b_b: 2, b_c_c: 3})
  %{"a" => 1, "b" => %{"a" => 1, "b" => 2, "c_c" => 3}}

  iex> RichardBurton.Codec.nest(%{a: 1, b_a: 1, b_b: 2, b_c_c_c: 3})
  %{"a" => 1, "b" => %{"a" => 1, "b" => 2, "c_c_c" => 3}}

  iex> RichardBurton.Codec.nest(%{a: 1, b_a: 1, b_b: 2, b_c_c: 3, b_c_d: 4})
  %{"a" => 1, "b" => %{"a" => 1, "b" => 2, "c" => %{"c" => 3, "d" => 4}}}