RichardBurton.Util (richard_burton v0.1.0)

Utilities for general purpose

Functions

create_fingerprint(data)

deep_merge_maps(map1, map2)

stringify_keys(s)

Given an enumerable with atom keys, return the enumerable with string keys instead. Works for keyword lists, key-value tuples, maps and structs. When an ordinary list is given, it stringify the keys of its elements. Does not traverse structures other than ordinary lists.

## Examples

  iex> RichardBurton.Util.stringify_keys(%{})
  %{}

  iex> RichardBurton.Util.stringify_keys(%{"key" => "value"})
  %{"key" => "value"}

  iex> RichardBurton.Util.stringify_keys({"key", "value"})
  {"key", "value"}

  iex> RichardBurton.Util.stringify_keys([])
  []

  iex> RichardBurton.Util.stringify_keys({:key, "value"})
  {"key", "value"}

  iex> RichardBurton.Util.stringify_keys({:key, :value})
  {"key", :value}

  iex> RichardBurton.Util.stringify_keys([key1: "value", key2: 1234])
  [{"key1", "value"}, {"key2", 1234}]

  iex> RichardBurton.Util.stringify_keys([key: :value])
  [{"key", :value}]

  iex> RichardBurton.Util.stringify_keys([{:key1, "value"}, {:key2, 1234}])
  [{"key1", "value"}, {"key2", 1234}]

  iex> RichardBurton.Util.stringify_keys([{:key1, "value"}, {"key2", 1234}])
  [{"key1", "value"}, {"key2", 1234}]

  iex> RichardBurton.Util.stringify_keys(%{key1: "value", key2: 1234})
  %{"key1" => "value", "key2" => 1234}

  iex> RichardBurton.Util.stringify_keys(%{key: :value})
  %{"key" => :value}

  iex> RichardBurton.Util.stringify_keys(%{:key1 => "value", "key2" => 1234})
  %{"key1" => "value", "key2" => 1234}

  iex> RichardBurton.Util.stringify_keys([{:key1, "value"}, %{key2: 1234}])
  [{"key1", "value"}, %{"key2" => 1234}]

  iex> RichardBurton.Util.stringify_keys(%{key: {:key, "value"}})
  %{"key" => {:key, "value"}}